+++ components/Notifications/ThankNotificationListing.tsx
@@ -136,7 +136,6 @@ export default function ThanksContentListing({
             className="rounded-2xl"
             width="40"
             height="40"
-            href={undefined}
             onClick={(e: React.MouseEvent) => {
               e.preventDefault(); // prevent outer link trigger
               e.stopPropagation();